backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.0.1"
description = ""
authors = ["idk"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
sqlalchemy = "^2.0.31"
fastapi = "^0.111.0"
python-dotenv = "^1.0.1"
pydantic-settings = "^2.3.3"
asyncpg = "^0.29.0"
loguru = "^0.7.2"
alembic = "^1.13.1"
passlib = "^1.7.4"
python-jose = "^3.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
